requirements.txt
google-adk>=0.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
requests>=2.31.0
